feat(versions): Switch package versions display to generic list
Instead having separate place to be able to review package version, this
will provide all information into the same location, for both core and
plugin packages.

This will depends on a Provider update which have to
be inherited by all SeAT package provider in order to be able to provide
proper information and implement auto-discovery.

This SeAT provider
should be able to return enough information to be able to build some
kind of marketplace and make plugin installation easier.