fix(search): Correct asset search query to use proper array (#90)
Bug:
When using the search as a user with affiliations for all corps and all chars, the search ended up not showing any results, while as a superuser they did.
a closer look at the query involved, and how it is build showed that it was doing a orWhere with a field name and an array. When checking if a fields value is within an array the use of orWhereIn is required, to ensure the array is treated as a list for an IN check, rather than as a string/int value for a comparison.

this is also in line with how similar queries are build in other parts of the search.