fix(assets): ensure in-space asset can be displayed
since ESI assets endpoint is not providing proper dataset, we need to extract some information from other tables in certain conditions (when root asset is in-space).

For example, when you have a ship in-space, like a Titan, it will not be listed into /characters/{character_id}/assets endpoint.

However, you can retrieve information from /characters/{character_id}/location (for the system name) and /character/{character_id}/ship (for the current ship, including the asset ID to which relay subsequent items)